requirements.txt
biopython
numpy
blast
matplotlib
scipy
pandas
mdtraj
mafft
pyrosetta
scrapy
ete2
